Rename registry extension to MdastImportExtension, add MdastExtension bundle
The split extensions are now symmetric: MdastImportExtension owns the
compiled registry and parsing, MdastExportExtension owns serialization,
and MdastExtension is a convenience bundle that depends on both for
simpler configuration when an editor wants both directions.

Feature extensions contribute their rules to MdastImportExtension
(unchanged mechanics, renamed target), and MdastShortcutsExtension
reads the registry from it.
