+++ packages/lexical-mdast/src/MdastStream.ts
@@ -64,7 +64,7 @@ function contentStartOffset(node: AnyMdastNode, line: string): number {
  * is no second, divergent set of regular expressions to keep in sync.
  *
  * It is constructed from the {@link CompiledMdast} registry assembled by
- * {@link MdastExtension}, so it stays in lock-step with whatever feature
+ * {@link MdastImportExtension}, so it stays in lock-step with whatever feature
  * extensions are enabled — including the inline construct types and trigger
  * characters contributed via `inlineShortcutTypes` / `inlineShortcutTriggers`.
  */
