+++ packages/lexical-mdast/src/compile.ts
@@ -6,13 +6,13 @@
  *
  */
 
-import type {MdastConfig} from './MdastExtension';
+import type {MdastConfig} from './MdastImportExtension';
 import type {CompiledMdast} from './types';
 
 /**
  * Compiles the raw contribution arrays held in {@link MdastConfig} into the
  * indexed registry used at runtime. Rules earlier in the arrays win for a
- * given node `type`; since {@link MdastExtension}'s `mergeConfig` prepends the
+ * given node `type`; since {@link MdastImportExtension}'s `mergeConfig` prepends the
  * rules contributed by extensions merged later (closer to the editor root),
  * those higher-priority rules take precedence — mirroring the dispatch order
  * of `@lexical/html`'s `DOMImportExtension`.
